Add nowplaying.resynth1943.net
It's basically a bookmarklet creator adapted from Mastodon Toot Bookmarklet.
This one automatically appends #nowplaying and removes the '- YouTube' from
the document title. This was made really quickly, but it works very well for the 
music-sharing crowd on the fediverse. This isn't a replacement for a scrobbler
which would automatically do this, but it's something :-)